requirements.txt
docling==2.40.0
docling_core==2.40.0
fastapi==0.116.1
langchain==0.3.26
langchain_community==0.3.27
langchain_core==0.3.68
langchain_openai==0.3.27
langchain_qdrant==0.2.0
langgraph==0.5.2
matplotlib==3.10.3
networkx==3.5
numpy==2.3.1
opencv_python_headless==4.12.0.88
pydantic==2.11.7
pydantic_settings==2.10.1
python-dotenv==1.1.1
qdrant_client==1.14.3
Requests==2.32.4
seaborn==0.13.2
sentence_transformers==5.0.0
uvicorn==0.35.0
Werkzeug==3.1.3
